% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_createCon.R
\name{dh_createCon}
\alias{dh_createCon}
\title{Create a database connection object}
\usage{
dh_createCon(connection_string)
}
\arguments{
\item{connection_string}{The name of the database to connect to. There are only two options: "postgres" or "cockroach".}
}
\value{
A DBI database object
}
\description{
Create a database connection object
}
\examples{
dh_createCon(connection_string = "cockroach")
}
