% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_getQuery.R
\name{dh_getQuery}
\alias{dh_getQuery}
\title{Query data from a database}
\usage{
dh_getQuery(connection, query, glue_params = NULL)
}
\arguments{
\item{connection}{A database connection object created by using the \code{dh_createCon} function.}

\item{query}{Reference to a SQL file stored in the \code{queries} directory, or a SQL query string.}

\item{glue_params}{Reference to any parameters that should be replaced in a SQL file.}
}
\value{
A dataframe
}
\description{
Query data from a database
}
