% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_ingestData.R
\name{dh_ingestData}
\alias{dh_ingestData}
\title{Write data to a database}
\usage{
dh_ingestData(connection, df, schema, table, append = TRUE)
}
\arguments{
\item{connection}{A database connection object created by using the \code{dh_createCon} function.}

\item{df}{A dataframe to ingest into database}

\item{schema}{The schema name of the database}

\item{table}{The table name of the database}

\item{append}{Whether or now to append df to an existing database object. Defaults to true, this means if the table doesn't exist it will be created.}
}
\value{
Nothing. A message is printed to signal success or failure.
}
\description{
Write data to a database
}
